refactor(features): split into framework-pure core + Granit.Http.Features + .Wolverine (#2367)
Granit.Features declared FrameworkReference Microsoft.AspNetCore.App, which is
transitive — so every dependent (including the now framework-pure
Granit.RateLimiting core) silently pulled in ASP.NET. Mirrors the #2364/#2366
RateLimiting split:

- Granit.Features (core, framework-pure): resolution, cache, IFeatureChecker,
  IFeatureLimitGuard, definitions, value types, exceptions, diagnostics, events,
  plans. Drops the ASP.NET FrameworkReference. Now lists Microsoft.Extensions.*
  as explicit transitive deps instead of relying on the ASP.NET shared framework.
- Granit.Http.Features (new): Minimal-API RequiresFeatureEndpointFilter +
  .RequiresFeature() extension. GranitHttpFeaturesModule [DependsOn(core)].
- Granit.Features.Wolverine (new, no WolverineFx — convention-discovered, no
  Wolverine types): pure RequiresFeatureAttribute + RequiresFeatureMiddleware.
  GranitFeaturesWolverineModule [DependsOn(core)].

Breaking (acceptable pre-1.0): RequiresFeatureAttribute loses IFilterFactory
(MVC controller auto-enforcement dropped — Granit is Minimal-API-first, zero
production usage) and moves to Granit.Features.Wolverine.Attributes; the
Minimal-API filter moves to Granit.Http.Features.AspNetCore; single module ->
three modules.

Tests split into Granit.Http.Features.Tests + Granit.Features.Wolverine.Tests
(the MVC RequiresFeatureFilter test is deleted). Granit.slnx, test-shards.json
(+regen .slnf), ArchitectureTests refs, lockfiles, and code index updated.
Architecture tests green (235). No THIRD-PARTY-NOTICES change.

Unblocks #2364's follow-up: with both RateLimiting and Features framework-pure,
Granit.AI.Extraction.StackExchangeRedis can migrate onto the RateLimiting core's
Redis counter store and delete its duplicated Lua.
