+++ .github/shard-filters/api-data.slnf
@@ -48,6 +48,7 @@
       "src/Granit.Http.Cookies/Granit.Http.Cookies.csproj",
       "src/Granit.Http.Cors/Granit.Http.Cors.csproj",
       "src/Granit.Http.ExceptionHandling/Granit.Http.ExceptionHandling.csproj",
+      "src/Granit.Http.Features/Granit.Http.Features.csproj",
       "src/Granit.Http.Idempotency/Granit.Http.Idempotency.csproj",
       "src/Granit.Http.ODataExposure/Granit.Http.ODataExposure.csproj",
       "src/Granit.Http.OutputCaching.StackExchangeRedis/Granit.Http.OutputCaching.StackExchangeRedis.csproj",
@@ -125,6 +126,7 @@
       "tests/Granit.Http.Cookies.Tests/Granit.Http.Cookies.Tests.csproj",
       "tests/Granit.Http.Cors.Tests/Granit.Http.Cors.Tests.csproj",
       "tests/Granit.Http.ExceptionHandling.Tests/Granit.Http.ExceptionHandling.Tests.csproj",
+      "tests/Granit.Http.Features.Tests/Granit.Http.Features.Tests.csproj",
       "tests/Granit.Http.Idempotency.Tests/Granit.Http.Idempotency.Tests.csproj",
       "tests/Granit.Http.ODataExposure.Tests/Granit.Http.ODataExposure.Tests.csproj",
       "tests/Granit.Http.OutputCaching.StackExchangeRedis.Tests/Granit.Http.OutputCaching.StackExchangeRedis.Tests.csproj",
