+++ .github/shard-filters/infrastructure.slnf
@@ -25,6 +25,7 @@
       "src/Granit.Events/Granit.Events.csproj",
       "src/Granit.Features.Endpoints/Granit.Features.Endpoints.csproj",
       "src/Granit.Features.EntityFrameworkCore/Granit.Features.EntityFrameworkCore.csproj",
+      "src/Granit.Features.Wolverine/Granit.Features.Wolverine.csproj",
       "src/Granit.Features/Granit.Features.csproj",
       "src/Granit.Guids/Granit.Guids.csproj",
       "src/Granit.Html.AngleSharp/Granit.Html.AngleSharp.csproj",
@@ -120,6 +121,7 @@
       "tests/Granit.Features.Endpoints.Tests/Granit.Features.Endpoints.Tests.csproj",
       "tests/Granit.Features.EntityFrameworkCore.Tests/Granit.Features.EntityFrameworkCore.Tests.csproj",
       "tests/Granit.Features.Tests/Granit.Features.Tests.csproj",
+      "tests/Granit.Features.Wolverine.Tests/Granit.Features.Wolverine.Tests.csproj",
       "tests/Granit.Html.AngleSharp.Tests/Granit.Html.AngleSharp.Tests.csproj",
       "tests/Granit.Html.Tests/Granit.Html.Tests.csproj",
       "tests/Granit.Localization.Endpoints.Tests/Granit.Localization.Endpoints.Tests.csproj",
