+++ .github/shard-filters/src-only.slnf
@@ -89,6 +89,7 @@
       "src/Granit.Events/Granit.Events.csproj",
       "src/Granit.Features.Endpoints/Granit.Features.Endpoints.csproj",
       "src/Granit.Features.EntityFrameworkCore/Granit.Features.EntityFrameworkCore.csproj",
+      "src/Granit.Features.Wolverine/Granit.Features.Wolverine.csproj",
       "src/Granit.Features/Granit.Features.csproj",
       "src/Granit.Guids/Granit.Guids.csproj",
       "src/Granit.Html.AngleSharp/Granit.Html.AngleSharp.csproj",
@@ -102,6 +103,7 @@
       "src/Granit.Http.Cookies/Granit.Http.Cookies.csproj",
       "src/Granit.Http.Cors/Granit.Http.Cors.csproj",
       "src/Granit.Http.ExceptionHandling/Granit.Http.ExceptionHandling.csproj",
+      "src/Granit.Http.Features/Granit.Http.Features.csproj",
       "src/Granit.Http.Idempotency/Granit.Http.Idempotency.csproj",
       "src/Granit.Http.ODataExposure/Granit.Http.ODataExposure.csproj",
       "src/Granit.Http.OutputCaching.StackExchangeRedis/Granit.Http.OutputCaching.StackExchangeRedis.csproj",
