+++ .github/test-shards.json
@@ -128,6 +128,7 @@
         "tests/Granit.Http.Cookies.Tests",
         "tests/Granit.Http.Cors.Tests",
         "tests/Granit.Http.ExceptionHandling.Tests",
+        "tests/Granit.Http.Features.Tests",
         "tests/Granit.Http.Idempotency.Tests",
         "tests/Granit.Http.ODataExposure.Tests",
         "tests/Granit.Http.OutputCaching.StackExchangeRedis.Tests",
@@ -172,6 +173,7 @@
         "tests/Granit.Features.Endpoints.Tests",
         "tests/Granit.Features.EntityFrameworkCore.Tests",
         "tests/Granit.Features.Tests",
+        "tests/Granit.Features.Wolverine.Tests",
         "tests/Granit.Localization.Endpoints.Tests",
         "tests/Granit.Localization.EntityFrameworkCore.Tests",
         "tests/Granit.Localization.SourceGenerator.Tests",
