+++ .slnf/application.slnf
@@ -27,6 +27,7 @@
       "src/Granit.Events/Granit.Events.csproj",
       "src/Granit.Features.Endpoints/Granit.Features.Endpoints.csproj",
       "src/Granit.Features.EntityFrameworkCore/Granit.Features.EntityFrameworkCore.csproj",
+      "src/Granit.Features.Wolverine/Granit.Features.Wolverine.csproj",
       "src/Granit.Features/Granit.Features.csproj",
       "src/Granit.Guids/Granit.Guids.csproj",
       "src/Granit.Http.ApiDocumentation/Granit.Http.ApiDocumentation.csproj",
@@ -91,6 +92,7 @@
       "tests/Granit.Features.Endpoints.Tests/Granit.Features.Endpoints.Tests.csproj",
       "tests/Granit.Features.EntityFrameworkCore.Tests/Granit.Features.EntityFrameworkCore.Tests.csproj",
       "tests/Granit.Features.Tests/Granit.Features.Tests.csproj",
+      "tests/Granit.Features.Wolverine.Tests/Granit.Features.Wolverine.Tests.csproj",
       "tests/Granit.Localization.AI.Tests/Granit.Localization.AI.Tests.csproj",
       "tests/Granit.Localization.Endpoints.Tests/Granit.Localization.Endpoints.Tests.csproj",
       "tests/Granit.Localization.EntityFrameworkCore.Tests/Granit.Localization.EntityFrameworkCore.Tests.csproj",
