+++ .slnf/framework-only.slnf
@@ -57,6 +57,7 @@
       "src/Granit.Events/Granit.Events.csproj",
       "src/Granit.Features.Endpoints/Granit.Features.Endpoints.csproj",
       "src/Granit.Features.EntityFrameworkCore/Granit.Features.EntityFrameworkCore.csproj",
+      "src/Granit.Features.Wolverine/Granit.Features.Wolverine.csproj",
       "src/Granit.Features/Granit.Features.csproj",
       "src/Granit.Guids/Granit.Guids.csproj",
       "src/Granit.Html.AngleSharp/Granit.Html.AngleSharp.csproj",
@@ -68,6 +69,7 @@
       "src/Granit.Http.Cookies/Granit.Http.Cookies.csproj",
       "src/Granit.Http.Cors/Granit.Http.Cors.csproj",
       "src/Granit.Http.ExceptionHandling/Granit.Http.ExceptionHandling.csproj",
+      "src/Granit.Http.Features/Granit.Http.Features.csproj",
       "src/Granit.Http.Idempotency/Granit.Http.Idempotency.csproj",
       "src/Granit.Http.ODataExposure/Granit.Http.ODataExposure.csproj",
       "src/Granit.Http.OutputCaching.StackExchangeRedis/Granit.Http.OutputCaching.StackExchangeRedis.csproj",
@@ -247,6 +249,7 @@
       "tests/Granit.Features.Endpoints.Tests/Granit.Features.Endpoints.Tests.csproj",
       "tests/Granit.Features.EntityFrameworkCore.Tests/Granit.Features.EntityFrameworkCore.Tests.csproj",
       "tests/Granit.Features.Tests/Granit.Features.Tests.csproj",
+      "tests/Granit.Features.Wolverine.Tests/Granit.Features.Wolverine.Tests.csproj",
       "tests/Granit.Guids.Tests/Granit.Guids.Tests.csproj",
       "tests/Granit.Html.AngleSharp.Tests/Granit.Html.AngleSharp.Tests.csproj",
       "tests/Granit.Html.Tests/Granit.Html.Tests.csproj",
@@ -256,6 +259,7 @@
       "tests/Granit.Http.Bulkhead.Tests/Granit.Http.Bulkhead.Tests.csproj",
       "tests/Granit.Http.Cors.Tests/Granit.Http.Cors.Tests.csproj",
       "tests/Granit.Http.ExceptionHandling.Tests/Granit.Http.ExceptionHandling.Tests.csproj",
+      "tests/Granit.Http.Features.Tests/Granit.Http.Features.Tests.csproj",
       "tests/Granit.Http.Idempotency.Tests/Granit.Http.Idempotency.Tests.csproj",
       "tests/Granit.Http.ODataExposure.Tests.Integration/Granit.Http.ODataExposure.Tests.Integration.csproj",
       "tests/Granit.Http.ODataExposure.Tests/Granit.Http.ODataExposure.Tests.csproj",
