+++ .slnf/platform.slnf
@@ -35,6 +35,7 @@
       "src/Granit.Http.Bulkhead/Granit.Http.Bulkhead.csproj",
       "src/Granit.Http.Cors/Granit.Http.Cors.csproj",
       "src/Granit.Http.ExceptionHandling/Granit.Http.ExceptionHandling.csproj",
+      "src/Granit.Http.Features/Granit.Http.Features.csproj",
       "src/Granit.Http.Idempotency/Granit.Http.Idempotency.csproj",
       "src/Granit.Http.ODataExposure/Granit.Http.ODataExposure.csproj",
       "src/Granit.Http.OutputCaching.StackExchangeRedis/Granit.Http.OutputCaching.StackExchangeRedis.csproj",
@@ -154,6 +155,7 @@
       "tests/Granit.Http.Bulkhead.Tests/Granit.Http.Bulkhead.Tests.csproj",
       "tests/Granit.Http.Cors.Tests/Granit.Http.Cors.Tests.csproj",
       "tests/Granit.Http.ExceptionHandling.Tests/Granit.Http.ExceptionHandling.Tests.csproj",
+      "tests/Granit.Http.Features.Tests/Granit.Http.Features.Tests.csproj",
       "tests/Granit.Http.Idempotency.Tests/Granit.Http.Idempotency.Tests.csproj",
       "tests/Granit.Http.ODataExposure.Tests.Integration/Granit.Http.ODataExposure.Tests.Integration.csproj",
       "tests/Granit.Http.ODataExposure.Tests/Granit.Http.ODataExposure.Tests.csproj",
