+++ tests/Granit.Http.Features.Tests/Granit.Http.Features.Tests.csproj
@@ -6,6 +6,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="xunit.v3" />
     <PackageReference Include="xunit.runner.visualstudio" />
@@ -15,7 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Granit.Features\Granit.Features.csproj" />
+    <ProjectReference Include="..\..\src\Granit.Http.Features\Granit.Http.Features.csproj" />
   </ItemGroup>
 
 </Project>
